Refactor types.ts to add XboxGamepadParams type
